Various: 1) doNextOne eliminated from WFT 2) operator= eliminated from GeometryDirection 3) operator= eliminated from VectorWithOffsets 4) Possible bug corrected in operator() for FermionSign

+++ src/Engine/FermionSign.h
@@ -119,7 +119,7 @@ namespace Dmrg {
 				io.read(signs_,"#FERMIONICSIGN");
 			}
 			
-			int operator()(size_t i,size_t f) const
+			int operator()(size_t i,int f) const
 			{
 				return (signs_[i]) ? f : 1;
 			}
